@@ -4,9 +4,7 @@ from warnings import warn
 
 from openpyxl.descriptors.serialisable import Serialisable
 from openpyxl.descriptors import (
-    Alias,
     Typed,
-    Sequence
 )
 from openpyxl.descriptors.sequence import NestedSequence
 from openpyxl.descriptors.excel import ExtensionList
@@ -23,15 +21,13 @@ from .fills import Fill
 from .fonts import Font
 from .numbers import (
     NumberFormatList,
+    BUILTIN_FORMATS,
+    BUILTIN_FORMATS_MAX_SIZE,
     BUILTIN_FORMATS_REVERSE,
     is_date_format,
     builtin_format_code
 )
-from .alignment import Alignment
-from .protection import Protection
 from .named_styles import (
-    NamedStyle,
-    _NamedCellStyle,
     _NamedCellStyleList
 )
 from .cell_style import CellStyle, CellStyleList
@@ -128,8 +124,12 @@ class Stylesheet(Serialisable):
         named_style.font = self.fonts[xf.fontId]
         named_style.fill = self.fills[xf.fillId]
         named_style.border = self.borders[xf.borderId]
-        if xf.numFmtId in self.custom_formats:
-            named_style.number_format = self.custom_formats[xf.numFmtId]
+        if xf.numFmtId < BUILTIN_FORMATS_MAX_SIZE:
+            formats = BUILTIN_FORMATS
+        else:
+            formats = self.custom_formats
+        if xf.numFmtId in formats:
+            named_style.number_format = formats[xf.numFmtId]
         if xf.alignment:
             named_style.alignment = xf.alignment
         if xf.protection:
@@ -164,7 +164,7 @@ class Stylesheet(Serialisable):
                 if fmt in BUILTIN_FORMATS_REVERSE: # remove builtins
                     style.numFmtId = BUILTIN_FORMATS_REVERSE[fmt]
                 else:
-                    style.numFmtId = formats.add(fmt) + 164
+                    style.numFmtId = formats.add(fmt) + BUILTIN_FORMATS_MAX_SIZE
             else:
                 fmt = builtin_format_code(style.numFmtId)
             if is_date_format(fmt):
@@ -226,7 +226,7 @@ def write_stylesheet(wb):
 
     from .numbers import NumberFormat
     fmts = []
-    for idx, code in enumerate(wb._number_formats, 164):
+    for idx, code in enumerate(wb._number_formats, BUILTIN_FORMATS_MAX_SIZE):
         fmt = NumberFormat(idx, code)
         fmts.append(fmt)
 
